requirements.txt
Flask==3.0.0
Pillow==10.3.0
qrcode==7.3.1
Werkzeug==3.0.2